Removed the unwanted abstract class
